FormatFix - Removal
Arden and I found an issue with this script. We couldn't figure out what was going wrong, but essentially, the top half (or so) of the file was the original .vcf that got fed in, while the bottom half was the .vcf file after going through the format fixing. We tried to remove this part and SKAT still worked. I am keeping it out because if your dataset it too small, this throws an error. Up until this point, we didn't know there was an issue because our files were large enough.