HeaderLines - counting fix; Option 2 fix
The original package was counting the number of header lines in the input .vcf; however, this number changes after some of our interal processing. The number of header lines we should be counting is in the .vcf that gets passed into FormatFix.sh. I've gone through to try and fix this.

Also improved text prompts for clarity, and consolidated some variables.

Also fixed input parameters for Option 2, or else the script would've failed going into ExautomateBackEnd.